Fixed bug with transaction and cleaned code
