+++ kafka-manager-deployment.yaml
@@ -1,3 +1,34 @@
+kind: PersistentVolume
+apiVersion: v1
+metadata:
+  name: kafka-manager-pv
+  labels:
+    type: local
+    app: kafka-manager
+spec:
+  storageClassName: manual
+  capacity:
+    storage: 1Gi
+  accessModes:
+    - ReadWriteOnce
+  hostPath:
+    path: "/mnt/data/kafka-manager"
+---
+kind: PersistentVolumeClaim
+apiVersion: v1
+metadata:
+  name: kafka-manager-pvc
+  namespace: kafka
+  labels:
+    app: kafka-manager
+spec:
+  storageClassName: manual
+  accessModes:
+    - ReadWriteOnce
+  resources:
+    requests:
+      storage: 1Gi
+---
 apiVersion: apps/v1
 kind: Deployment
 metadata:
@@ -24,6 +55,13 @@ spec:
           command:
             - ./bin/kafka-manager
             - -Dhttp.port=80
+          volumeMounts:
+            - name: kafka-storage
+              mountPath: /conf
+      volumes:
+        - name: kafka-storage
+          persistentVolumeClaim:
+            claimName: kafka-manager-pvc
 ---
 kind: Service
 apiVersion: v1
