+++ src/main/java/com/stepuro/payment/order/service/TransferEntityService.java
@@ -1,0 +1,8 @@
+package com.stepuro.payment.order.service;
+
+import com.stepuro.payment.order.api.dto.PaymentOrderEntityDto;
+
+public interface TransferEntityService {
+    PaymentOrderEntityDto createRestClientPayment(PaymentOrderEntityDto paymentOrderEntityDto);
+    PaymentOrderEntityDto createPayment(PaymentOrderEntityDto paymentOrderEntityDto);
+}
