+++ src/main/resources/application.properties
@@ -26,3 +26,8 @@ webclient.timeout.millis=5000
 spring.kafka.bootstrap-servers=kafka-service:9092
 kafka.topic-name.transfer-amount=transferAmount
 kafka.group.id=paymentOrderGroup
+
+## Liveness and readiness probes
+management.endpoint.health.probes.enabled=true
+management.health.livenessState.enabled=true
+management.health.readinessState.enabled=true
